Services: Moved ApplicationChoiceSubmission
