+++ spec/services/candidate_interface/application_choice_submission_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe CandidateInterface::ContinuousApplications::ApplicationChoiceSubmission do
+RSpec.describe CandidateInterface::ApplicationChoiceSubmission do
   include ActionView::Helpers::UrlHelper
   include GovukLinkHelper
   include GovukVisuallyHiddenHelper
